react router package installed and updated App.js according to routing
