+++ package.json
@@ -10,6 +10,7 @@
     "gh-pages": "^4.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.0",
     "react-scripts": "5.0.1",
     "sass": "^1.54.9",
     "web-vitals": "^2.1.4"
