+++ public/index.html
@@ -27,7 +27,7 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Lato:wght@400&display=swap" rel="stylesheet">
   <!-- Google Fonts -->
-  <title>Crown</title>
+  <title>Crown Clothing</title>
 </head>
 
 <body>
